app now has toastr messages

+++ src/app/display-items/display-items/display-items.component.ts
@@ -2,6 +2,7 @@ import { DataService } from './../../services/data.service';
 import { ItemComponent } from './../../item/item.component';
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-display-items',
@@ -13,7 +14,10 @@ export class DisplayItemsComponent implements OnInit {
   items: ItemComponent[];
   item: ItemComponent;
 
-  constructor(private dataService: DataService, private router: Router, private activatedRoute: ActivatedRoute) { }
+  constructor(private dataService: DataService,
+              private router: Router,
+              private activatedRoute: ActivatedRoute,
+              private toastr: ToastrService) { }
 
   ngOnInit(): void {
     this.getData();
@@ -34,9 +38,15 @@ export class DisplayItemsComponent implements OnInit {
   }
 
   deleteItem(id: number): void {
-    this.dataService.deleteItem(id).subscribe((data: void) => {
-      let index: number = this.items.findIndex(item => item.id === id);
-      this.items.splice(index, 1);
-    });
+    if(confirm("Are you sure you want to delete this item?")) {
+      this.dataService.deleteItem(id).subscribe((data: void) => {
+        let index: number = this.items.findIndex(item => item.id === id);
+        this.items.splice(index, 1);
+      });
+      this.toastr.error("Item deleted!");
+    }
+    setTimeout(function(){
+      location.reload();
+    },1500);
   }
 }
