+++ src/app/item-details/item-details.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ItemComponent } from '../item/item.component';
 import { DataService } from '../services/data.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-item-details',
@@ -13,7 +14,10 @@ export class ItemDetailsComponent implements OnInit {
   item: ItemComponent;
   pageTitle: string = 'Item Details'
 
-  constructor(private dataService: DataService, private route: ActivatedRoute, private router: Router) { }
+  constructor(private dataService: DataService,
+              private route: ActivatedRoute,
+              private router: Router,
+              private toastr: ToastrService) { }
 
   ngOnInit(): void {
     const param = this.route.snapshot.paramMap.get('id');
@@ -30,10 +34,13 @@ export class ItemDetailsComponent implements OnInit {
   }
 
   deleteItem(id: number): void {
-    this.dataService.deleteItem(id).subscribe((data: void) => {
-      // let index: number = this.items.findIndex(item => item.id === id);
-      // this.items.splice(index, 1);
-    });
+    if(confirm("Are you sure you want to delete this item?")) {
+      this.dataService.deleteItem(id).subscribe((data: void) => {
+        // let index: number = this.items.findIndex(item => item.id === id);
+        // this.items.splice(index, 1);
+      });
+      this.toastr.error("Item deleted!");
+    }
   }
 
   onBack(): void {
